Revert "End-to-end tests for AssignIPV3"
