+++ vpc/service/branch_enis.go
@@ -17,7 +17,6 @@ import (
 	"github.com/Netflix/titus-executor/vpc/service/vpcerrors"
 	"github.com/Netflix/titus-executor/vpc/tracehelpers"
 	"github.com/aws/aws-sdk-go/aws"
-	request "github.com/aws/aws-sdk-go/aws/request"
 	"github.com/aws/aws-sdk-go/service/ec2"
 	"github.com/google/uuid"
 	"github.com/lib/pq"
@@ -1201,7 +1200,7 @@ WHERE subnets.subnet_id = $1
 		},
 	}
 
-	output, err := session.CreateNetworkInterface(ctx, createNetworkInterfaceInput, request.Option(nil))
+	output, err := session.CreateNetworkInterface(ctx, createNetworkInterfaceInput)
 	logger.G(ctx).WithField("createNetworkInterfaceInput", createNetworkInterfaceInput).Debug("Creating Branch ENI")
 	if err != nil {
 		err = errors.Wrap(err, "Cannot create branch network interface")
